prompt_master_backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9  # PostgreSQL适配器
python-dotenv==1.0.1  # 环境变量管理